AT-6113 adding common objects to the init file on PAgeObject
