+++ uitests/framework/page_objects/login_page.py
@@ -4,7 +4,7 @@ from selenium import webdriver
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.wait import WebDriverWait
 
-from uitests.framework.page_objects.common_methods import JediCommonMethods
+from . import PageObjectMethods
 
 
 class Login:
