+++ uitests/framework/page_objects/new_portfolio_page.py
@@ -1,13 +1,13 @@
 import random
 import string
 
+from selenium import webdriver
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.by import By
-from selenium import webdriver
 from selenium.webdriver.support.wait import WebDriverWait
-
-from uitests.framework.page_objects.common_methods import JediCommonMethods
 from selenium.webdriver.support import expected_conditions as EC
+
+from . import PageObjectMethods
 
 
 class AddNewPortfolioPages:
