+++ uitests/framework/page_objects/settings_page.py
@@ -1,4 +1,4 @@
-from uitests.framework.page_objects.common_methods import JediCommonMethods
+from . import PageObjectMethods
 
 
 class SettingsPages:
