+++ uitests/framework/utilities/read_properties.py
@@ -2,7 +2,6 @@ import configparser
 import os
 
 config = configparser.RawConfigParser()
-config.read(".//Configurations//config.ini")
 
 
 class ReadConfig:
@@ -15,13 +14,3 @@ class ReadConfig:
     def getLoginLocalURL():
         url2 = os.getenv("loginURL")
         return url2
-
-    @staticmethod
-    def getUserName():
-        username = config.get("common items", "username")
-        return username
-
-    @staticmethod
-    def getPassword():
-        password = config.get("common items", "password")
-        return password
